Make google chart fit entire page [1/1]
[Problem]
Google chart is squashed at the top of the page because the <html> and <body> tags do not have their width and height
set properly. This works in "quirks" mode when <!DOCTYPE> is not specified, but this is not good practice.

[Solution]
Set the flights CSS file to have html and body have full size of the browser screen with width and height equal to 100%.