Fix SCSS layouts for flights/home controller [1/1]
[Problem]
The SCSS layouts defined for the application for the home view and the flight view did not jive well.
Home and photos navigation bar links did not have assets.

[Solution]
Fix the sassy css files to work well between the two controllers while keeping as much of the same application scss
definitions as possible.
Create assets for the home and photos links.