Add links for home/photos images [1/1]
[Problem]
The home/photos images do nothing causing an inability to use the hover/selected states of the navigation targets.

[Solution]
Create 4-state assets for home/photos (selected x hover) and separate controller targets for "home" and "photos". The
data for which navigation element is selected is passed from the controller into the view.