Issue-1: Migrate flight information to rails db [1/1]
[Problem]
Rails needs to query the database for flight information and then dynamically create the Google Chart javascript

[Solution]
Add navigation for flight queries to the navigation bar which posts back the flight query id when retreiving the
corresponding web page.

When hitting a page to display the chart, the flight information is looked up using the id and then the Google Chart JS
is generated at runtime.