QPX may only contain an single origin and destination
If multiple origins or destinations may be used such as where there are multiple airports in a region e.g. in Seoul
there is ICN and GMP, then these requests must occur in separate API calls. Round-trips may be in the same call, but
need to be in chronological order.

More parameters were added to the request object.